@@ -3,23 +3,23 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "framer-motion": "^4.1.17",
+    "@testing-library/jest-dom": "^6.4.2",
+    "@testing-library/react": "^15.0.0",
+    "@testing-library/user-event": "^14.5.2",
+    "framer-motion": "^11.0.27",
     "immer": ">=9.0.6",
     "medium-zoom": "^1.0.6",
     "nth-check": ">=2.0.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-icons": "^4.3.1",
-    "react-intersection-observer": "^8.32.2",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "4.0.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-icons": "^5.0.1",
+    "react-intersection-observer": "^9.8.2",
+    "react-router-dom": "^6.22.3",
+    "react-scripts": "5.0.1",
     "react-scroll": "^1.8.4",
     "react-transition-group": "^4.4.2",
-    "styled-components": "^5.3.3",
-    "web-vitals": "^1.1.2"
+    "styled-components": "^6.1.8",
+    "web-vitals": "^3.5.2"
   },
   "scripts": {
     "start": "react-scripts start",
