@@ -5,7 +5,7 @@ import Photo from '../../images/Photo.webp'
 
 const transition = {
     duration: 0.3,
-    ease: [0.6, 0.01, -0.05, 0.9]
+    // ease: [0.6, 0.01, -0.05, 0.9]
 }
 
 const appearUp = {
