@@ -3,7 +3,7 @@ import { HeroContainer, HeroContent, HeroH1, HeroSpan, LinkContainer, HeroLink }
 
 const transition = {
     duration: 0.3,
-    ease: [0.6, 0.01, -0.05, 0.9]
+    // ease: [0.6, 0.01, -0.05, 0.9]
 }
 
 const appearUp = {
