@@ -6,7 +6,7 @@ import { useAnimation } from 'framer-motion'
 
 const transition = {
     duration: 0.3,
-    ease: [0.6, 0.01, -0.05, 0.9]
+    // ease: [0.6, 0.01, -0.05, 0.9]
 }
 
 const appearUp = {
