@@ -13,8 +13,8 @@ export const SidebarContainer = styled.aside`
     top: 0;
     left: 0;
     transition: 0.3s ease-in-out;
-    opacity: ${({ isOpen }) => (isOpen ? '100%' : '0')};
-    top: ${({ isOpen }) => (isOpen ? '0' : '-100%')};
+    opacity: ${({ $isOpen }) => ($isOpen ? '100%' : '0')};
+    top: ${({ $isOpen }) => ($isOpen ? '0' : '-100%')};
 `
 
 export const CloseIcon = styled(FaTimes)`
